refactor: infra/database 패키지 생성
