+++ src/main/java/com/hhpl/concertreserve/infra/database/concert/ReservationJpaRepository.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.infra.concert;
+package com.hhpl.concertreserve.infra.database.concert;
 
 import com.hhpl.concertreserve.domain.concert.model.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
