+++ src/main/java/com/hhpl/concertreserve/infra/database/payment/PaymentRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.infra.payment;
+package com.hhpl.concertreserve.infra.database.payment;
 
 import com.hhpl.concertreserve.domain.payment.PaymentRepository;
 import com.hhpl.concertreserve.domain.payment.model.Payment;
