+++ src/test/java/com/hhpl/concertreserve/application/ConcertFacadeIntegrationTest.java
@@ -10,9 +10,9 @@ import com.hhpl.concertreserve.domain.concert.model.Schedule;
 import com.hhpl.concertreserve.domain.concert.model.Seat;
 import com.hhpl.concertreserve.domain.concert.type.ReservationStatus;
 import com.hhpl.concertreserve.domain.concert.type.SeatStatus;
-import com.hhpl.concertreserve.infra.concert.ConcertJpaRepository;
-import com.hhpl.concertreserve.infra.concert.ScheduleJpaRepository;
-import com.hhpl.concertreserve.infra.concert.SeatJpaRepository;
+import com.hhpl.concertreserve.infra.database.concert.ConcertJpaRepository;
+import com.hhpl.concertreserve.infra.database.concert.ScheduleJpaRepository;
+import com.hhpl.concertreserve.infra.database.concert.SeatJpaRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
