+++ src/test/java/com/hhpl/concertreserve/application/PaymentFacadeIntegrationTest.java
@@ -12,12 +12,12 @@ import com.hhpl.concertreserve.domain.concert.type.SeatStatus;
 import com.hhpl.concertreserve.domain.error.CoreException;
 import com.hhpl.concertreserve.domain.user.model.Point;
 import com.hhpl.concertreserve.domain.waitingqueue.model.WaitingQueue;
-import com.hhpl.concertreserve.infra.concert.ConcertJpaRepository;
-import com.hhpl.concertreserve.infra.concert.ReservationJpaRepository;
-import com.hhpl.concertreserve.infra.concert.ScheduleJpaRepository;
-import com.hhpl.concertreserve.infra.concert.SeatJpaRepository;
-import com.hhpl.concertreserve.infra.user.PointJpaRepository;
-import com.hhpl.concertreserve.infra.waitingqueue.WaitingQueueJpaRepository;
+import com.hhpl.concertreserve.infra.database.concert.ConcertJpaRepository;
+import com.hhpl.concertreserve.infra.database.concert.ReservationJpaRepository;
+import com.hhpl.concertreserve.infra.database.concert.ScheduleJpaRepository;
+import com.hhpl.concertreserve.infra.database.concert.SeatJpaRepository;
+import com.hhpl.concertreserve.infra.database.user.PointJpaRepository;
+import com.hhpl.concertreserve.infra.database.waitingqueue.WaitingQueueJpaRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
